pyproject.toml
[project]
name = "senior-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "python-jose>=3.5.0",
    "pytest>=9.0.1",
]
